requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.5
djangorestframework==3.14.0
idna==3.4
mssql-django==1.2
mysql==0.0.3
mysqlclient==2.1.1
numpy==1.24.1
pandas==1.5.3
Pillow==9.4.0
pyodbc==4.0.35
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
